more path/name moves
