+++ update-server/otupdate/common/config.py
@@ -1,5 +1,5 @@
 """
-otupdate.common.config: Handlers for reading update server configuration
+otupdate.buildroot.config: Handlers for reading update server configuration
 """
 
 import os
@@ -13,20 +13,17 @@ from . import constants
 
 LOG = logging.getLogger(__name__)
 
-DEFAULT_CERT_PATH = "/etc/opentrons-robot-signing-key.crt"
-REQUIRED_DATA = [
-    ("signature_required", bool, True),
-    ("download_storage_path", str, "/var/lib/otupdate/downloads"),
-    ("update_cert_path", str, DEFAULT_CERT_PATH),
-]
-DEFAULT_PATH = "/var/lib/otupdate/config.json"
-PATH_ENVIRONMENT_VARIABLE = "OTUPDATE_CONFIG_PATH"
-CONFIG_VARNAME = constants.APP_VARIABLE_PREFIX + "config"
+DEFAULT_CERT_PATH = '/etc/opentrons-robot-signing-key.crt'
+REQUIRED_DATA = [('signature_required', bool, True),
+                 ('download_storage_path', str, '/var/lib/otupdate/downloads'),
+                 ('update_cert_path', str, DEFAULT_CERT_PATH)]
+DEFAULT_PATH = '/var/lib/otupdate/config.json'
+PATH_ENVIRONMENT_VARIABLE = 'OTUPDATE_CONFIG_PATH'
+CONFIG_VARNAME = constants.APP_VARIABLE_PREFIX + 'config'
 
 
 class Config(NamedTuple):
-    """Configuration elements for the update server"""
-
+    """ Configuration elements for the update server """
     signature_required: bool
     #: Whether the system requires updates to be signed
     download_storage_path: str
@@ -43,7 +40,7 @@ def config_from_request(req: Request) -> Config:
 
 def _ensure_load(path: str) -> Optional[Mapping[str, Any]]:
     try:
-        contents = open(path, "r").read()
+        contents = open(path, 'r').read()
     except OSError:
         LOG.exception("Couldn't load config file, defaulting")
         return None
@@ -59,7 +56,7 @@ def _ensure_load(path: str) -> Optional[Mapping[str, Any]]:
 
 
 def _ensure_values(data: Mapping[str, Any]) -> Tuple[Dict[str, Any], bool]:
-    """Make sure we have appropriate keys and say if we should write"""
+    """ Make sure we have appropriate keys and say if we should write """
     to_return = {}
     should_write = False
     for keyname, typekind, default in REQUIRED_DATA:
@@ -70,8 +67,7 @@ def _ensure_values(data: Mapping[str, Any]) -> Tuple[Dict[str, Any], bool]:
         elif not isinstance(data[keyname], typekind):
             LOG.warning(
                 f"Config value {keyname} was {type(data[keyname])} not"
-                f" {typekind}, defaulted to {default}"
-            )
+                f" {typekind}, defaulted to {default}")
             to_return[keyname] = default
             should_write = True
         else:
@@ -88,14 +84,13 @@ def load_from_path(path: str) -> Config:
     if not data:
         data = {}
     values, should_write = _ensure_values(data)
-    values.update({"path": path})
+    values.update({'path': path})
     config = Config(**values)
     if config.signature_required:
         if not os.path.exists(config.update_cert_path):
             LOG.warning(
                 f"No signing cert is present in {config.update_cert_path}, "
-                "code signature checking disabled"
-            )
+                "code signature checking disabled")
             config = config._replace(signature_required=False)
             config = config._replace(update_cert_path=DEFAULT_CERT_PATH)
     if should_write:
@@ -104,9 +99,10 @@ def load_from_path(path: str) -> Config:
 
 
 def _get_path(args_path: Optional[str]) -> str:
-    """Find the valid path from args then env then default"""
+    """ Find the valid path from args then env then default """
     env_path = os.getenv(PATH_ENVIRONMENT_VARIABLE)
-    for path, source in ((args_path, "arg"), (env_path, "env")):
+    for path, source in ((args_path, 'arg'),
+                         (env_path, 'env')):
         if not path:
             LOG.debug(f"config.load: skipping {source} (path None)")
             continue
@@ -128,10 +124,11 @@ def save_to_path(path: str, config: Config) -> None:
     Save the config file to a specific path (not what's in the config)
     """
     LOG.debug(f"Saving config to {path}")
-    with open(path, "w") as cf:
-        cf.write(json.dumps({k: v for k, v in config._asdict().items() if k != "path"}))
+    with open(path, 'w') as cf:
+        cf.write(json.dumps({k: v for k, v in config._asdict().items()
+                             if k != 'path'}))
 
 
 def save(config: Config) -> None:
-    """Save the config element back to wherever it was loaded"""
+    """ Save the config element back to wherever it was loaded """
     save_to_path(config.path, config)
