+++ update-server/otupdate/common/name_management.py
@@ -1,5 +1,5 @@
 """
-otupdate.common.name_management: functions for managing machine names
+otupdate.buildroot.name_management: functions for managing machine names
 
 The robot has two names associated with it:
 
@@ -33,14 +33,11 @@ try:
     import dbus
 
     class DBusState:
-        """Bundle of state for dbus"""
-
-        def __init__(
-            self,
-            bus: dbus.SystemBus,
-            server: dbus.Interface,
-            entrygroup: dbus.Interface,
-        ) -> None:
+        """ Bundle of state for dbus """
+        def __init__(self,
+                     bus: dbus.SystemBus,
+                     server: dbus.Interface,
+                     entrygroup: dbus.Interface) -> None:
             """
             Build the state bundle.
 
@@ -65,29 +62,29 @@ try:
         global _BUS_STATE
         if not _BUS_STATE:
             bus = dbus.SystemBus()
-            server_obj = bus.get_object("org.freedesktop.Avahi", "/")
-            server_if = dbus.Interface(server_obj, "org.freedesktop.Avahi.Server")
+            server_obj = bus.get_object('org.freedesktop.Avahi', '/')
+            server_if = dbus.Interface(
+                server_obj, 'org.freedesktop.Avahi.Server')
             entrygroup_path = server_if.EntryGroupNew()
-            entrygroup_obj = bus.get_object("org.freedesktop.Avahi", entrygroup_path)
+            entrygroup_obj = bus.get_object(
+                'org.freedesktop.Avahi', entrygroup_path)
             entrygroup_if = dbus.Interface(
-                entrygroup_obj, "org.freedesktop.Avahi.EntryGroup"
-            )
+                entrygroup_obj, 'org.freedesktop.Avahi.EntryGroup')
             _BUS_STATE = DBusState(bus, server_if, entrygroup_if)
 
         _BUS_STATE.entrygroup.Reset()
         hostname = _BUS_STATE.server.GetHostName()
         domainname = _BUS_STATE.server.GetDomainName()
         _BUS_STATE.entrygroup.AddService(
-            dbus.Int32(-1),  # avahi.IF_UNSPEC
-            dbus.Int32(-1),  # avahi.PROTO_UNSPEC
-            dbus.UInt32(0),  # flags
-            name,  # sname
-            "_http._tcp",  # stype
-            domainname,  # sdomain (.local)
-            f"{hostname}.{domainname}",  # shost (hostname.local)
-            dbus.UInt16(31950),  # port
-            dbus.Array([], signature="ay"),
-        )
+            dbus.Int32(-1),                  # avahi.IF_UNSPEC
+            dbus.Int32(-1),                  # avahi.PROTO_UNSPEC
+            dbus.UInt32(0),                  # flags
+            name,                            # sname
+            '_http._tcp',                    # stype
+            domainname,                      # sdomain (.local)
+            f'{hostname}.{domainname}',      # shost (hostname.local)
+            dbus.UInt16(31950),              # port
+            dbus.Array([], signature='ay'))
         _BUS_STATE.entrygroup.Commit()
 
 except ImportError:
@@ -101,7 +98,7 @@ _BUS_LOCK = asyncio.Lock()
 
 
 def _get_hostname() -> str:
-    """Get a good value for the system hostname.
+    """ Get a good value for the system hostname.
 
     The hostname is loaded from, in order of preference,
 
@@ -109,11 +106,12 @@ def _get_hostname() -> str:
       not empty, and not the default
     - the systemd-generated machine-id, which changes at every boot.
     """
-    if os.path.exists("/var/serial"):
-        serial = open("/var/serial").read().strip()
+    if os.path.exists('/var/serial'):
+        serial = open('/var/serial').read().strip()
         if serial:
             LOG.info("Using serial for hostname")
-            hn = "".join([c for c in urllib.parse.quote(serial, safe="") if c != "%"])
+            hn = ''.join([c for c in urllib.parse.quote(serial, safe='')
+                          if c != '%'])
             if hn != serial:
                 LOG.warning(f"Reencoded serial to {hn}")
             return hn
@@ -123,7 +121,7 @@ def _get_hostname() -> str:
     else:
         LOG.info("Using machine-id for hostname: no /var/serial")
 
-    return open("/etc/machine-id").read().strip()[:6]
+    return open('/etc/machine-id').read().strip()[:6]
 
 
 async def setup_hostname() -> str:
@@ -139,67 +137,58 @@ async def setup_hostname() -> str:
     :returns: the hostname
     """
     hostname = _get_hostname()
-    with open("/etc/hostname", "w") as ehn:
-        ehn.write(f"{hostname}\n")
+    with open('/etc/hostname', 'w') as ehn:
+        ehn.write(f'{hostname}\n')
 
     # First, we run hostnamed which will set the transient hostname
     # and loaded static hostname from the value we just wrote to
     # /etc/hostname
     LOG.debug("Setting hostname")
     proc = await asyncio.create_subprocess_exec(
-        "hostname",
-        hostname,
-        stdout=asyncio.subprocess.PIPE,
-        stderr=asyncio.subprocess.PIPE,
-    )
+        'hostname', '-F', '/etc/hostname',
+        stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE)
     stdout, stderr = await proc.communicate()
     ret = proc.returncode
     if ret != 0:
         LOG.error(
-            f"Error starting hostname: {ret} " f"stdout: {stdout!r} stderr: {stderr!r}"
-        )
+            f'Error starting hostname: {ret} '
+            f'stdout: {stdout!r} stderr: {stderr!r}')
         raise RuntimeError("Couldn't run hostname")
 
     # Then, with the hostname set, we can restart avahi
     LOG.debug("Restarting avahi")
     proc = await asyncio.create_subprocess_exec(
-        "systemctl",
-        "restart",
-        "avahi-daemon",
-        stdout=asyncio.subprocess.PIPE,
-        stderr=asyncio.subprocess.PIPE,
-    )
+        'systemctl', 'restart', 'avahi-daemon',
+        stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE)
     stdout, stderr = await proc.communicate()
     ret = proc.returncode
     if ret != 0:
         LOG.error(
-            f"Error restarting avahi-daemon: {ret} "
-            f"stdout: {stdout!r} stderr: {stderr!r}"
-        )
+            f'Error restarting avahi-daemon: {ret} '
+            f'stdout: {stdout!r} stderr: {stderr!r}')
         raise RuntimeError("Error restarting avahi")
     LOG.debug("Updated hostname and restarted avahi OK")
     return hostname
 
 
 def _update_pretty_hostname(new_val: str):
-    """Write a new value for the pretty hostname.
+    """ Write a new value for the pretty hostname.
 
     :raises OSError: If the new value could not be written.
     """
     try:
-        with open("/etc/machine-info") as emi:
+        with open('/etc/machine-info') as emi:
             contents = emi.read()
     except OSError:
         LOG.exception("Couldn't read /etc/machine-info")
-        contents = ""
+        contents = ''
     new_contents = _rewrite_machine_info(contents, new_val)
-    with open("/etc/machine-info", "w") as emi:
+    with open('/etc/machine-info', 'w') as emi:
         emi.write(new_contents)
 
 
 def _rewrite_machine_info(
-    current_machine_info_contents: str, new_pretty_hostname: str
-) -> str:
+        current_machine_info_contents: str, new_pretty_hostname: str) -> str:
     """
     Return current_machine_info_contents - the full contents of
     /etc/machine-info - with the PRETTY_HOSTNAME=... line rewritten to refer
@@ -207,24 +196,24 @@ def _rewrite_machine_info(
     """
     current_lines = current_machine_info_contents.splitlines()
     preserved_lines = [
-        ln for ln in current_lines if not ln.startswith("PRETTY_HOSTNAME")
-    ]
-    new_lines = preserved_lines + [f"PRETTY_HOSTNAME={new_pretty_hostname}"]
-    new_contents = "\n".join(new_lines) + "\n"
+        ln for ln in current_lines if not ln.startswith('PRETTY_HOSTNAME')]
+    new_lines = preserved_lines + [f'PRETTY_HOSTNAME={new_pretty_hostname}']
+    new_contents = '\n'.join(new_lines) + '\n'
     return new_contents
 
 
-def get_name(default: str = "no name set"):
-    """Get the currently-configured name of the machine"""
+def get_name(default: str = 'no name set'):
+    """ Get the currently-configured name of the machine """
     try:
-        with open("/etc/machine-info") as emi:
+        with open('/etc/machine-info') as emi:
             contents = emi.read()
     except OSError:
-        LOG.exception("Couldn't read /etc/machine-info")
-        contents = ""
-    for line in contents.split("\n"):
-        if line.startswith("PRETTY_HOSTNAME="):
-            return "=".join(line.split("=")[1:])
+        LOG.exception(
+            "Couldn't read /etc/machine-info")
+        contents = ''
+    for line in contents.split('\n'):
+        if line.startswith('PRETTY_HOSTNAME='):
+            return '='.join(line.split('=')[1:])
     LOG.warning(f"No PRETTY_HOSTNAME in {contents}, defaulting to {default}")
     try:
         _update_pretty_hostname(default)
@@ -254,13 +243,12 @@ async def set_name(name: str = None) -> str:
 
     async with _BUS_LOCK:
         await asyncio.get_event_loop().run_in_executor(
-            None, _set_name_future, checked_name
-        )
+            None, _set_name_future, checked_name)
     return checked_name
 
 
 async def set_name_endpoint(request: web.Request) -> web.Response:
-    """Set the name of the robot.
+    """ Set the name of the robot.
 
     Request with POST /server/name {"name": new_name}
     Responds with 200 OK {"hostname": new_name, "prettyname": pretty_name}
@@ -273,20 +261,23 @@ async def set_name_endpoint(request: web.Request) -> web.Response:
     """
 
     def build_400(msg: str) -> web.Response:
-        return web.json_response(data={"message": msg}, status=400)
+        return web.json_response(
+            data={'message': msg},
+            status=400)
 
     body = await request.json()
-    if "name" not in body or not isinstance(body["name"], str):
+    if 'name' not in body or not isinstance(body['name'], str):
         return build_400('Body has no "name" key with a string')
 
-    new_name = await set_name(body["name"])
+    new_name = await set_name(body['name'])
     request.app[DEVICE_NAME_VARNAME] = new_name
 
-    return web.json_response(data={"name": new_name}, status=200)
+    return web.json_response(data={'name': new_name},
+                             status=200)
 
 
 async def get_name_endpoint(request: web.Request) -> web.Response:
-    """Get the name of the robot.
+    """ Get the name of the robot.
 
     This information is also accessible in /server/update/health, but this
     endpoint provides symmetry with POST /server/name.
@@ -294,5 +285,5 @@ async def get_name_endpoint(request: web.Request) -> web.Response:
     GET /server/name -> 200 OK, {'name': robot name}
     """
     return web.json_response(
-        data={"name": request.app[DEVICE_NAME_VARNAME]}, status=200
-    )
+        data={'name': request.app[DEVICE_NAME_VARNAME]},
+        status=200)
